Slight refactor of createReadStream
